# swagger/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API documentation",
        "title": "Go Boilerplate API",
        "contact": {
            "name": "Abdan Hafidz",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/events/{event_slug}/exam": {
            "get": {
                "description": "Retrieve a list of exams associated with a specific event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exam Event"
                ],
                "summary": "List Exams by Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug",
                        "name": "event_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-array_models_Exam"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{event_slug}/exam/{attempt_id}/answer_question": {
            "post": {
                "description": "Submit an answer for a specific question in an exam attempt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exam Event"
                ],
                "summary": "Answer Exam Event Question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug",
                        "name": "event_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exam Attempt ID",
                        "name": "attempt_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Answer Exam Event Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AnswerExamEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-dto_AnswerExamEventRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{event_slug}/exam/{exam_slug}/attempt": {
            "get": {
                "description": "Start an attempt for a specific exam in an event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exam Event"
                ],
                "summary": "Attempt Exam Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug",
                        "name": "event_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exam Slug",
                        "name": "exam_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-models_ExamEventAttempt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/options/create": {
            "post": {
                "description": "Create multiple options in bulk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Option"
                ],
                "summary": "Create Bulk Options",
                "parameters": [
                    {
                        "description": "Bulk Options Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.OptionsRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/options/list/{slug}": {
            "get": {
                "description": "Retrieve option details using its slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Option"
                ],
                "summary": "Get Option by Slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Option Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-models_Options"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/options/region/cities": {
            "get": {
                "description": "Retrieve a list of cities within a specific province",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Region"
                ],
                "summary": "List Cities by Province",
                "parameters": [
                    {
                        "description": "Province ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "province_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-array_models_RegionCity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/options/region/provinces": {
            "get": {
                "description": "Retrieve a list of all provinces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Region"
                ],
                "summary": "List Provinces",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-array_models_RegionProvince"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/options/region/seed-cities": {
            "post": {
                "description": "Seed multiple cities into the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Region"
                ],
                "summary": "Seed Cities",
                "parameters": [
                    {
                        "description": "Seed Cities Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RegionCity"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-array_models_RegionCity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/options/region/seed-provinces": {
            "post": {
                "description": "Seed multiple provinces into the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Region"
                ],
                "summary": "Seed Provinces",
                "parameters": [
                    {
                        "description": "Seed Provinces Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RegionProvince"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AnswerExamEventRequest": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.OptionsRequest": {
            "type": "object",
            "required": [
                "option_name",
                "option_values"
            ],
            "properties": {
                "option_name": {
                    "type": "string"
                },
                "option_values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.SuccessResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-array_models_Exam": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Exam"
                    }
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessResponse-array_models_RegionCity": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RegionCity"
                    }
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessResponse-array_models_RegionProvince": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RegionProvince"
                    }
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-dto_AnswerExamEventRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AnswerExamEventRequest"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-models_ExamEventAttempt": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.ExamEventAttempt"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessResponse-models_Options": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Options"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_detail_completed": {
                    "type": "boolean"
                },
                "is_email_verified": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.Events": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "end_event": {
                    "type": "string"
                },
                "event_code": {
                    "type": "string"
                },
                "id_event": {
                    "type": "string"
                },
                "img_banner": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "overview": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "register_status": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "start_event": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Exam": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/models.ExamConfiguration"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id_exam": {
                    "type": "string"
                },
                "proctoring": {
                    "$ref": "#/definitions/models.ExamProctoring"
                },
                "randomize": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.ExamConfiguration": {
            "type": "object",
            "properties": {
                "allow_retake": {
                    "type": "boolean"
                },
                "allow_review": {
                    "type": "boolean"
                },
                "enable_timer": {
                    "type": "boolean"
                },
                "id_exam": {
                    "type": "string"
                },
                "id_result": {
                    "type": "string"
                }
            }
        },
        "models.ExamEventAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "exam_attempt": {
                    "$ref": "#/definitions/models.ExamEventAttempt"
                },
                "id": {
                    "type": "string"
                },
                "id_attempt": {
                    "type": "string"
                },
                "id_question": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ExamEventAttempt": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExamEventAnswer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "due_at": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/models.Events"
                },
                "exam": {
                    "$ref": "#/definitions/models.Exam"
                },
                "id_account": {
                    "type": "string"
                },
                "id_attempt": {
                    "type": "string"
                },
                "id_event": {
                    "type": "string"
                },
                "id_exam": {
                    "type": "string"
                },
                "mark": {
                    "type": "number"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Questions"
                    }
                },
                "remaining_time": {
                    "type": "integer"
                },
                "submitted": {
                    "type": "boolean"
                }
            }
        },
        "models.ExamProctoring": {
            "type": "object",
            "properties": {
                "disable_copy_paste": {
                    "type": "boolean"
                },
                "enable_exam_browser": {
                    "type": "boolean"
                },
                "enable_eye_tracking": {
                    "type": "boolean"
                },
                "enable_full_screen": {
                    "type": "boolean"
                },
                "enable_tab_block": {
                    "type": "boolean"
                },
                "enable_vad": {
                    "type": "boolean"
                },
                "enable_webcam": {
                    "type": "boolean"
                },
                "id_exam": {
                    "type": "string"
                },
                "id_result": {
                    "type": "string"
                }
            }
        },
...
        "models.Options": {
            "type": "object",
            "properties": {
                "option_category": {
                    "$ref": "#/definitions/models.OptionCategory"
                },
                "option_values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OptionValues"
                    }
                }
            }
        },
...
        "models.RegionCity": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "full_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "province_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.RegionProvince": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}